Fix get_mon_map() for octopus and later
The "ceph mon_status" command seems to have disappeared on octopus and
later, and is replaced by "ceph quorum_status". This changes the
get_mon_map() function to detect the underlying ceph version and do the
right thing.

Note that the fix is actually in charm-helpers, and this has been
manually synced into the charm [1].

Note: also picked pyparsing pin to allow tests to run.

Cherry-picked from master.

[1] https://github.com/juju/charm-helpers/pull/659

Change-Id: I59cf6fc19cf2a91b0aef37059cdb0ed37379b5cb
Closes-Bug: #1951094