Backport the latest charms.ceph changes to stable/pacific
This patchset applies the latest changes in charms.ceph to pacific,
to work around an issue due to a command renaming. In addition, to
make the unit tests pass, every import that involved 'mock' had to
be rewritten to 'unittest.mock'.

Change-Id: Ic093ae4c7462c496b94f271ddf5413515070deab
(cherry picked from commit 444f91f559bad93ad7091871715257504942bf25)